Wayland: Do not destroy the registry in registerGlobals
We still want to receive globals events after registering all globals.
For example, we want to modified when new monitors are added or existing
monitors and inputs are removed. Therefore, we must destroy the registry
in the Wayland destructor.

This fixes a bug where the bar was not drawn on hotplugged monitors.